brain_tumor_detector/requirements.txt
torch==2.4.1
torchvision==0.19.1
timm==1.0.7
streamlit==1.38.0
Pillow==10.4.0
numpy==2.1.1
pandas==2.2.2
scikit-learn==1.5.2
matplotlib==3.9.2
seaborn==0.13.2
albumentations==1.4.18
tensorboard==2.17.1
opencv-python==4.10.0.84
tqdm==4.66.5
